Factor TensorInfo and DataId into a separate file
