+++ tfjs-core/src/engine_test.ts
@@ -20,7 +20,7 @@ import {ENGINE} from './engine';
 import * as tf from './index';
 import {KernelFunc} from './index';
 import {ALL_ENVS, describeWithFlags, TestKernelBackend} from './jasmine_util';
-import {TensorInfo} from './kernel_registry';
+import { TensorInfo } from './tensor_info';
 import {Tensor} from './tensor';
 import {expectArraysClose} from './test_util';
 import {BackendValues, DataType} from './types';
