+++ tfjs-core/src/kernel_registry.ts
@@ -19,33 +19,32 @@ import {getGlobal} from './global_util';
 import * as log from './log';
 import {NamedGradientMap} from './tape';
 import {Tensor} from './tensor';
-import {DataType, RecursiveArray} from './types';
+import {TensorInfo} from './tensor_info';
+import {RecursiveArray} from './types';
 
 const kernelRegistry =
-    getGlobal('kernelRegistry', () => new Map<`${string}_${string}`,
-              KernelConfig>());
+  getGlobal('kernelRegistry', () => new Map<`${string}_${string}`,
+    KernelConfig>());
 const gradRegistry =
-    getGlobal('gradRegistry', () => new Map<string, GradConfig>());
-
-export type DataId = object;
+  getGlobal('gradRegistry', () => new Map<string, GradConfig>());
 
 type AttributeValue =
-    number|number[]|boolean|boolean[]|string|string[]|NamedAttrMap;
+  number | number[] | boolean | boolean[] | string | string[] | NamedAttrMap;
 
 /** These are extra non-tensor/primitive params passed to kernel functions. */
-export type Attribute = AttributeValue|RecursiveArray<AttributeValue>;
+export type Attribute = AttributeValue | RecursiveArray<AttributeValue>;
 
 /** Specifies the code to run when executing a kernel. */
 export type KernelFunc = (params: {
   inputs: NamedTensorInfoMap,
   backend: {},
   attrs?: NamedAttrMap,
-}) => TensorInfo|TensorInfo[];
+}) => TensorInfo | TensorInfo[];
 
 /** The function to run when computing a gradient during backprop. */
 export type GradFunc =
-    (dy: Tensor|Tensor[], saved: Tensor[], attrs: NamedAttrMap) =>
-        NamedGradientMap;
+  (dy: Tensor | Tensor[], saved: Tensor[], attrs: NamedAttrMap) =>
+    NamedGradientMap;
 
 /** Function that gets called after the backend initializes. */
 export type KernelSetupFunc = (backend: {}) => void;
@@ -70,13 +69,6 @@ export interface GradConfig {
   saveAllInputs?: boolean;
   outputsToSave?: boolean[];
   gradFunc: GradFunc;
-}
-
-/** Holds metadata for a given tensor. */
-export interface TensorInfo {
-  dataId: DataId;
-  shape: number[];
-  dtype: DataType;
 }
 
 export interface NamedTensorInfoMap {
