+++ tfjs-core/src/kernel_registry_test.ts
@@ -18,7 +18,8 @@
 import * as tf from './index';
 import {KernelBackend} from './index';
 import {ALL_ENVS, describeWithFlags, TestEnv} from './jasmine_util';
-import {KernelFunc, TensorInfo} from './kernel_registry';
+import { KernelFunc } from './kernel_registry';
+import { TensorInfo } from './tensor_info';
 import {expectArraysClose} from './test_util';
 
 describeWithFlags('kernel_registry', ALL_ENVS, (testEnv: TestEnv) => {
