+++ tfjs-core/src/ops/gather_nd_util.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  * =============================================================================
  */
-import {TensorInfo} from '../kernel_registry';
+import { TensorInfo } from '../tensor_info';
 import {computeStrides, sizeFromShape} from '../util';
 
 /**
