+++ tfjs-core/src/tensor.ts
@@ -19,7 +19,7 @@
 /// <reference types="@webgpu/types/dist" />
 
 import {getGlobal} from './global_util';
-import {TensorInfo} from './kernel_registry';
+import {TensorInfo, DataId} from './tensor_info';
 import {tensorToString} from './tensor_format';
 import {ArrayMap, BackendValues, DataType, DataTypeMap, DataValues, NumericDataType, Rank, ShapeMap, SingleValueMap, TypedArray} from './types';
 import * as util from './util';
@@ -238,16 +238,6 @@ export function setOpHandler(handler: OpHandler) {
 export function setDeprecationWarningFn(fn: (msg: string) => void) {
   deprecationWarningFn = fn;
 }
-
-/**
- * We wrap data id since we use weak map to avoid memory leaks.
- * Since we have our own memory management, we have a reference counter
- * mapping a tensor to its data, so there is always a pointer (even if that
- * data is otherwise garbage collectable).
- * See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/
- * Global_Objects/WeakMap
- */
-export type DataId = object;  // object instead of {} to force non-primitive.
 
 // Declare this namespace to make Tensor class augmentation work in google3.
 export declare namespace Tensor {}
