+++ tfjs-core/src/tensor_info.ts
@@ -1,0 +1,35 @@
+/**
+ * @license
+ * Copyright 2022 Google LLC.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {DataType} from './types';
+
+/**
+ * We wrap data id since we use weak map to avoid memory leaks.
+ * Since we have our own memory management, we have a reference counter
+ * mapping a tensor to its data, so there is always a pointer (even if that
+ * data is otherwise garbage collectable).
+ * See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/
+ * Global_Objects/WeakMap
+ */
+export type DataId = object;  // object instead of {} to force non-primitive.
+
+/** Holds metadata for a given tensor. */
+export interface TensorInfo {
+  dataId: DataId;
+  shape: number[];
+  dtype: DataType;
+}
